chore: added appropriate error handling for delete endpoint

+++ backend/controllers/song.controller.js
@@ -50,13 +50,16 @@ export const updateSong = async (req, res) => {
 
 export const deleteSong = async (req, res) => {
     const { id } = req.params;
-    console.log("id:", id);
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return  res.status(404).json({ success:false, message: "Song Not Found" });
+    }
 
     try {
         await Song.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Song Deleted" });
     } catch (error) {
         console.error("Error in deleting the song:", error.message);
-        res.status(404).json({ success: false, message: "Song Not Found" });
+        res.status(500).json({ success: false, message: "Server Error" });
     }
 };
